Removed connAck process buggy test.
The test "incorrect versions" was not exercising an failing path due
to mismatching versions, but was failing because of error in consensus
state verification (ConsensusStateVerificationFailure). This test
is no longer useful since all versions are set to the default
and therefore always match, so there is no failing path to test.