Prevent race condition when updating the workers after config reload (#1173)
This PR fixes a bug where the supervisor would attempt to shutdown a worker
that had already been shutdown and replaced with a new worker. As the new worker was not
instructed to shutdown, the supervisor would hang waiting for it to exit.

In more details:

- Say we have a packet worker A for object O between ibc-0 and ibc-1
- After updating the config of ibc-0, we trigger a config reload
- The supervisor first shuts down worker A, and waits for its event loop to exit
- Worker A exits, and sends a `Stopped(O)` message to the supervisor
- The supervisor now spawns a new worker B for object O
- The supervisor now processes the `Stopped(O)` message previously sent by worker A,
   and waits for the worker associated with object O, ie. now worker B, to finish
- As worker B has never been instructed to shutdown, nor should it, the supervisor hangs
  waiting for it to exit.

This PR solves this issue by:
- introducing unique identifiers for each worker (currently an incrementing `u64`)
- sending the worker id alongside the object in the `Stopped` message
- only acting on worker `ID` when receiving a `Stopped(ID, OBJECT)` message

Additionally, the `SIGUSR1` signal will now also show the worker's ids when dumping the supervisor state.