Prevent overflow in fee amount computation (#1568)
Prevent a potential overflow when computing the fee for a tx to be sent to a chain which uses a denomination which high decimal places.

This is done by using infinite precision integers and rationals in the `mul_ceil` function. The resulting `BigInt` is then formatted as a string in the `Fee` proto.

Because the computation is now done will full precision, and because some decimal numbers cannot be represented precisely as a 64-bit IEEE-754 floating point number, the result of the computation may sometime change, eg. be rounded up to the next integer. This should not be a problem as the gas and fee computations are not fully precise anyway, given that they rely on the simulated gas.

---

* Fix clippy warning

* Compute fee amount using `BigInt`

* Propagate new signature of `mul_ceil` and check for overflow

* Add config validation step for `gas_adjustment`

* Add changelog entry

* Fix warning

* Fix unit tests

* Add unit test to demonstrate that the previous behavior would overflow when given high gas and price

* Fix typo in comment

* Turn debug assertions into normal ones