Adding support for base64 encoding when serializing
Go and prost have different ways of serializing ``Vec<u8>`` a.k.a.
``[]byte``. [1][2]
Go encodes the type in base64 format, while rust encodes it as a normal
vector of numbers.
If we want to give full cosmwasm support, in those serde
serialized/deserialized structures we need to properly encode/decode
arrays of bytes in a way that we can correctly communicate with Go.

This commit adds support for base64 encoding/decoding of ``Vec<u8>``
when serializing/deserialing fields of such type

[1] https://play.rust-lang.org/?version=stable&mode=debug&edition=2021&gist=18e810ab0a6795a1988a8b342a9b669b
[2] https://go.dev/play/p/gOm8rZRHbef