Remove channel version negotiation step during channel handshake
In IBC v2, relayers would query the `NegotiateAppVersion` endpoint before calling `ChanOpenTry`.
Applications would then need to verify that the passed in version was correct.

As of IBC v3, applications will perform this version negotiation during the channel handshake,
thus removing the need for the relayer to query the version.