Refactor `send_tx` methods into plain functions (#2044)
This PR attempts to rework the core logic of `send_tx` methods in `CosmosSdkChain` into plain functions. This is essential in helping us to solve the root problems in issues such as #2012, #2040, and supporting sending of transactions with fees in ICS29.

At this stage, I have managed to refactor `send_tx_with_account_sequence_retry` and its inner calls into plain functions, and simplify the data access a little.

Since this is a refactoring PR, no logic should be changed. The integration tests should help to ensure that nothing breaks with the refactoring.

---

* Refactoring send_tx

* WIP refactoring

* More code reorganization

* Split out functions from main cosmos.rs

* Use refactored code to send_tx

* Walk through send_tx_with_account_sequence

* Refactor code

* Reorder function arguments

* Refactor send_tx_with_account_sequence_retry into plain function

* Refactor account query functions

* Refactor query_tx

* Refactor wait_for_block_commits

* Turn wait_for_block_commits into simple loop

* Refactor send_messages_and_wait_commit

* Refactor send_messages_and_wait_check_tx

* Refactor sign_message

* Refactor gas config

* Move out query account module

* Reorganize types

* Remove pub const

* Simplify arguments

* Remove redundant account query function

* Refactor query status

* Introduce TransferTimeout abstraction

* Use prost::Message::encoded_len() to compute encoded message length

* Address review feedback

* Re-add missing comments

* Fix clippy error

* Remove check for both timeout height offset and duration being zero

* Do not set timeout height or time when input is zero