Do not retry clearing with same height forever (#2348)
## Description
`spawn_packet_cmd_worker` runs `handle_packet_cmd` passing a "command" that can be `NewBlock`.
The first thing it does is to try clearing packets. If this fails with ignorable error, the old code will retry immediately with same command and height. If the error persists (as in #2155) then the worker enters an infinite loop and new IBC events (coming via `IbcEvent` commands) are never processed.

I believe the infinite loop started with https://github.com/informalsystems/ibc-rs/pull/2238. Before that the worker would abort (due to `TaskError::Fatal(RunError::retry(e))`)

The propose change is to move to the next command even if previous has failed. The reasons are:
- it avoids the infinite loop
- the failed events will still be processed at next clear interval with a fresh height
- at different levels down in `handle_packet_cmd` there are retries mechanisms for MAX_RETRIES (current value hardcoded at 5). 

## Commits

* Do not retry clearing with same height forever

* Undo one-chain script changes

* Add changelog

* Reword changelog entry

* Remove dbg! statement

* Formatting

Co-authored-by: Romain Ruetschi <romain@informal.systems>