batch: fix size estimation for empty body
When the TxBody struct is completely empty (all fields have default
values), it is encoded as an empty array and the body_bytes field is
then also omitted from the encoding of TxRaw.

Add transaction size verification to the unit tests of batch_messages,
by creating encoded tx from the batched messages (using the maximum fee
to keep unit tests simple) and checking the size corresponds to the
`MaxTxSize` parameter.