Factor rpc_client out of TxConfig
There is already an HttpClient instantiated in the chain driver,
and it's better to move it out of TxConfig which is supposed to only
contain state relevant to transaction processing.

Also remove the duplication of transaction size limit parameters in
the batch module functions. The same parameters are passed via TxConfig.