Do not perform client update if consensus state already exists
On some chains there are a few active channels using the same connection
and therefore clients. This PR adds a check in the client update code and
not create the message if the consensus state at desired height is already present.
This helps with reducing relayer fees, state bloating, etc by eliminating redundant updates.