wip: make config an endpoint-specific enum
Work towards https://github.com/informalsystems/hermes/issues/3636

This turns out to be a very good place to start to discover requirements for
Hermes to operate in a more multi-chain way, since it allows discovering which
parts of the code access Cosmos-specific data to do Cosmos-specific logic, and
which parts are required for general relaying.

The only change to the existing config format so far is the change in semantics
of the `type` field, which now must be `CosmosSdk`. I did not preserve the
existing logic and tests that parsing different variations of casing and dashes
are supported, because I don't think this is particularly valuable to support.

So far, only the core `ibc-relayer` crate is updated; the cli is not, because
it's not necessarily clear how the cli should change to accomodate non-SDK
chains.