Make `ChainEndpoint::get_key` take `&self` instead of `&mut self` (#3732)
Updates the `get_key` method on the `ChainEndpoint` trait to take `&self` instead of `&mut self`, as it wasn't needed for self to be mutable. it's also expected that a function named as `get` doesn't mutate self. this also allows `get_key` to be used in methods that take `&self`.